Rename the components name
