+++ components/NpcLogo.vue
